Began Implementing ExpressionStatement, specifically the Expr_BinaryOp_Concat part.
Discovered a bug in the GitHub parser that causes weird parsing by my parser. Implemented a detector and fixer so my parser works even if the bug occurs.

Added and changed tests so they test the variable put in the TaintMap by the computeTaintFromInput functions also add taint to the variable itself.