Introduced TaintMap Class, and updated the tests and code to use this class.
Implemented A function which runs the tool against a Dat file, as well as resolving the data flow equations.

Fixed a bug in the parsing of phi functions, as they are delimited with a ", " not just a ","

Changed the TaintAnalyser as well as the blocks so that a work set is used to record which blocks need their taint status to be updated. Blocks now no longer need to remember if they need to be updated, this is done by the main function using a workset.